requirements.txt
numpy==2.4.2
pandas==3.0.1
pyarrow==23.0.1
plotly==6.5.2
requests==2.32.5
python-dotenv==1.2.1
scikit-learn==1.8.0
xgboost==3.2.0
scipy==1.17.1
openpyxl==3.1.5
lxml==6.0.2
